# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Excel Importer REST API",
    "version": "v1"
  },
  "paths": {
...
    "/api/excel-tool/export-html-table": {
      "post": {
        "operationId": "exportHtmlTable",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "sheetName": {
                    "type": "string"
                  },
                  "tableHtml": {
                    "format": "byte",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportHtmlTableResult"
                }
              }
            },
            "description": "Successful export"
          }
        },
        "summary": "Export html table as excel sheet",
        "tags": [
          "Excel Tool"
        ]
      }
    },
    "/api/excel-tool/exports/{exportId}/wait": {
      "get": {
        "operationId": "waitForExport",
        "parameters": [
          {
            "description": "ID of the export",
            "in": "path",
            "name": "exportId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Timeout in seconds",
            "in": "query",
            "name": "timeout",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "description": "default response"
          }
        },
        "summary": "Wait for export completion",
        "tags": [
          "Excel Tool"
        ]
      }
    },
    "/api/excel-tool/html-table-from-list": {
      "post": {
        "operationId": "getHTMLTable",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "array"
              }
            }
          },
          "description": "List of data to be converted into HTML table",
          "required": true
        },
        "responses": {
          "200": {
            "description": "HTML content"
          }
        },
        "summary": "Generate HTML table",
        "tags": [
          "Excel Tool"
        ]
      }
    },
    "/api/excel-tool/projects/{projectId}/attach-table": {
      "post": {
        "operationId": "attachTable",
        "parameters": [
          {
            "description": "Project ID",
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachTableParams"
              }
            }
          },
          "description": "Parameters to attach table",
          "required": true
        },
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "description": "default response"
          }
        },
        "summary": "Attach table to project",
        "tags": [
          "Excel Tool"
        ]
      }
    },
    "/api/import/jobs": {
      "get": {
        "operationId": "getAllImporterJobs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Import jobs statuses"
          }
        },
        "summary": "Returns all active import jobs statuses",
        "tags": [
          "Excel Processing"
        ]
      },
      "post": {
        "operationId": "startImportJob",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "mappingName": {
                    "default": "Default",
                    "type": "string"
                  },
                  "projectId": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Import process is started, job URI is returned in Location header"
          }
        },
        "summary": "Starts asynchronous import job",
        "tags": [
          "Excel Processing"
        ]
      }
    },
    "/api/import/jobs/{id}": {
      "get": {
        "operationId": "getImportJobStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobDetails"
                }
              }
            },
            "description": "Import job is still in progress"
          },
          "303": {
            "content": {
              "application/*": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobDetails"
                }
              }
            },
            "description": "Import job is finished successfully, Location header contains result URL"
          },
          "404": {
            "description": "Import job id is unknown"
          },
          "409": {
            "description": "Import job is failed or cancelled"
          }
        },
        "summary": "Returns import job status",
        "tags": [
          "Excel Processing"
        ]
      }
    },
    "/api/import/jobs/{id}/result": {
      "get": {
        "operationId": "getImportJobResult",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "Import result is ready"
          },
          "204": {
            "description": "Import job is still in progress"
          },
          "404": {
            "description": "Import job id is unknown"
          },
          "409": {
            "description": "Import job is failed, cancelled or result is unreachable"
          }
        },
        "summary": "Returns import job result",
        "tags": [
          "Excel Processing"
        ]
      }
    },
    "/api/projects/{projectId}/import": {
      "post": {
        "operationId": "importExcelSheet",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  },
                  "mappingName": {
                    "default": "Default",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            },
            "description": "Successful import"
          }
        },
        "summary": "Imports Excel sheet",
        "tags": [
          "Excel Processing"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttachTableParams": {
        "description": "Parameters required to attach a table to a project",
        "properties": {
          "fileName": {
            "description": "Name of the file to be created",
            "type": "string"
          },
          "fileTitle": {
            "description": "Title of the file to be displayed",
            "type": "string"
          },
          "htmlTable": {
            "description": "HTML representation of the table to attach",
            "type": "string"
          },
          "objectId": {
            "description": "Unique identifier of the object",
            "type": "string"
          },
          "objectType": {
            "description": "Type of the object (e.g., Work Item, Attachment)",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ExportHtmlTableResult": {
        "description": "Represents the result of a html table export operation",
        "properties": {
          "content": {
            "description": "Base64-encoded excel file content",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ImportJobDetails": {
        "description": "Details of the import job including status and error message if any",
        "properties": {
          "errorMessage": {
            "description": "Error message if the import failed",
            "type": "string"
          },
          "status": {
            "description": "Status of the import job",
            "enum": [
              "IN_PROGRESS",
              "SUCCESSFULLY_FINISHED",
              "FAILED",
              "CANCELLED"
            ],
            "example": "IN_PROGRESS",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ImportResult": {
        "description": "Represents the result of an Excel import operation",
        "properties": {
          "createdIds": {
            "description": "List of IDs that were created during the import process",
            "items": {
              "description": "List of IDs that were created during the import process",
              "type": "string"
            },
            "type": "array"
          },
          "log": {
            "description": "Text log generated during import process",
            "type": "string"
          },
          "skippedIds": {
            "description": "List of IDs that were skipped during the import process",
            "items": {
              "description": "List of IDs that were skipped during the import process",
              "type": "string"
            },
            "type": "array"
          },
          "unchangedIds": {
            "description": "List of IDs that were unchanged during the import process",
            "items": {
              "description": "List of IDs that were unchanged during the import process",
              "type": "string"
            },
            "type": "array"
          },
          "updatedIds": {
            "description": "List of IDs that were updated during the import process",
            "items": {
              "description": "List of IDs that were updated during the import process",
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    }
  }
}